requirements.txt
-e git+git://github.com/b1naryth1ef/disco.git#egg=disco-py
asn1crypto==0.24.0
cachetools==2.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
enum34==1.1.6
fonttools==3.25.0
futures==3.2.0
gevent==1.2.2
google-api-core==1.1.2
google-auth==1.4.1
google-cloud-texttospeech==0.1.0
googleapis-common-protos==1.5.3
greenlet==0.4.13
grpcio==1.11.0
holster==2.0.0
idna==2.6
ipaddress==1.0.22
pendulum==2.0.0
Pillow==5.1.0
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyNaCl==1.2.1
pyOpenSSL==17.5.0
python-dateutil==2.7.3
pytz==2018.4
pytzdata==2018.4
qrcode==6.0
requests==2.18.4
rsa==3.4.2
six==1.11.0
TTFQuery==1.0.5
typing==3.6.4
urllib3==1.22
Wand==0.4.4
websocket-client==0.44.0
